+++ app/src/main/java/com/example/calco/logic/persistent/entities/ProductImages.java
@@ -18,5 +18,5 @@ public class ProductImages {
 
     @NonNull
     @ColumnInfo(name = "product_id")
-    public int productId;
+    public long productId;
 }
